Add the possibility to exclude routes from CSRF
I've got a concrete use case : I'm building an app who receive "xml report' from an external webservice. 
The route which receive those calls is "www.myapp.com/service/report".
The webservice send a POST Request to my app, on this route, and I need to read the body of this request. But I cannot control the content of the request : it's external, it's not filled by myself.
Because of the POST method, the request is under the check of the CSRF-Guard and it's refused (doesn't contain any token).
Then, I would love to exclude that route from the CSRF-Guard. 

Usage : 
```
$csrfGuard = new \Slim\Extras\Middleware\CsrfGuard();
$csrfGuard->addExcludedRoutes( 'my-route-name' ); // you can also pass a array of route-name
$app->add($csrfGuard);
```

If you think that isn't "clean" to exclude your route where you declare the Middleware, you can make a Slim Singleton :
```
$csrfGuard = new \Slim\Extras\Middleware\CsrfGuard();
$app->container->singleton('csrf_guard_middleware', function($c) use ($csrfGuard) {
    return $csrfGuard;
});
$app->add($csrfGuard);
unset($csrfGuard);
```
And now, when you declare your routes, you can do :
```
$app->post('service/report', ..blabla... )->name('xml-report');
$app->csrf_guard_middleware->addExcludedRoutes( 'xml-report' ); // means that route doesn't need CSRF Guard
```